Backup/requirements.txt
line-bot-sdk
Flask
mysql-connector-python
google.generativeai
